factor out abstractions related to consensus/validity
