+++ src/models/blockchain/transaction/validity/typescripts_halt.rs
@@ -6,13 +6,14 @@ use serde::{Deserialize, Serialize};
 use triton_vm::prelude::{BFieldElement, Claim, NonDeterminism, Program};
 use twenty_first::shared_math::bfield_codec::BFieldCodec;
 
-use crate::models::blockchain::transaction::{
-    transaction_kernel::TransactionKernel,
-    utxo::{TypeScript, Utxo},
-    PrimitiveWitness,
+use crate::models::{
+    blockchain::transaction::{
+        transaction_kernel::TransactionKernel,
+        utxo::{TypeScript, Utxo},
+        TransactionPrimitiveWitness,
+    },
+    consensus::{ClaimSupport, SecretWitness, SupportedClaim, ValidationLogic},
 };
-
-use super::{ClaimSupport, SecretWitness, SupportedClaim, ValidationLogic};
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, GetSize, BFieldCodec)]
 pub struct TypeScriptHaltsWitness {
@@ -46,8 +47,12 @@ impl TypeScriptsHalt {
 }
 
 impl ValidationLogic<TypeScriptHaltsWitness> for TypeScriptsHalt {
+    type PrimitiveWitness = TransactionPrimitiveWitness;
+
+    type Kernel = TransactionKernel;
+
     fn new_from_primitive_witness(
-        primitive_witness: &PrimitiveWitness,
+        primitive_witness: &TransactionPrimitiveWitness,
         tx_kernel: &TransactionKernel,
     ) -> Self {
         let claim = Claim {
