+++ src/models/mod.rs
@@ -1,6 +1,7 @@
 pub mod big_array;
 pub mod blockchain;
 pub mod channel;
+pub mod consensus;
 pub mod database;
 pub mod peer;
 pub mod shared;
